refactor: add layout for login & leaderboard to fix responsive layout
